# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SEO Matrix - All-in-One SEO & Schema API",
    "description": "Ultra-fast micro-services for JSON-LD Schema generation, LSI keywords, Helpful Content audits, URL slugifying, readability, and meta inspection.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "https://seotoolkitapi.vercel.app",
      "description": "Production Server (Vercel)"
    }
  ],
  "paths": {
...
    "/v1/seo/slugify": {
      "post": {
        "summary": "Generate Slug",
        "description": "Generate clean, SEO-friendly, stop-word-free URL slug.",
        "operationId": "generate_slug_v1_seo_slugify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlugifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlugifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seo/extract-keywords": {
      "post": {
        "summary": "Extract Keywords",
        "description": "Extract top keywords and calculate keyword density percentage.",
        "operationId": "extract_keywords_v1_seo_extract_keywords_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeywordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seo/meta-inspect": {
      "post": {
        "summary": "Inspect Meta",
        "description": "Inspect HTML Meta Title, Meta Description, H1 tags, Canonical & OG tags.",
        "operationId": "inspect_meta_v1_seo_meta_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetaInspectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaInspectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seo/readability": {
      "post": {
        "summary": "Calculate Readability",
        "description": "Calculate Word Count, Character Count, Estimated Reading Time, and Difficulty Level.",
        "operationId": "calculate_readability_v1_seo_readability_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadabilityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seo/schema-generate": {
      "post": {
        "summary": "Generate Schema",
        "description": "Generate valid JSON-LD Structured Data (Schema.org) for Article, Product, or FAQ.",
        "operationId": "generate_schema_v1_seo_schema_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemaGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seo/lsi-keywords": {
      "post": {
        "summary": "Extract Lsi Keywords",
        "description": "Extract N-gram phrases (2-word and 3-word co-occurrences) for LSI Keyword Analysis.",
        "operationId": "extract_lsi_keywords_v1_seo_lsi_keywords_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LsiKeywordsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LsiKeywordsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seo/content-audit": {
      "post": {
        "summary": "Audit Content",
        "description": "Perform a heuristic Helpful Content audit (word count, paragraph structure, readability score).",
        "operationId": "audit_content_v1_seo_content_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentAuditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seo/link-check": {
      "post": {
        "summary": "Check Links",
        "description": "Extract and classify internal vs external links, anchor text, and nofollow attributes from HTML.",
        "operationId": "check_links_v1_seo_link_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/seo/robots-generator": {
      "post": {
        "summary": "Generate Robots",
        "description": "Generate clean robots.txt directives and Sitemap declaration.",
        "operationId": "generate_robots_v1_seo_robots_generator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RobotsGeneratorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotsGeneratorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeywordRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "example": "FastAPI is a modern, fast Web framework for Python. Python is great for AI."
          },
          "top_n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top N",
            "default": 5,
            "example": 5
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "KeywordRequest"
      },
      "KeywordResponse": {
        "properties": {
          "total_words": {
            "type": "integer",
            "title": "Total Words"
          },
          "top_keywords": {
            "items": {
              "$ref": "#/components/schemas/KeywordItem"
            },
            "type": "array",
            "title": "Top Keywords"
          }
        },
        "type": "object",
        "required": [
          "total_words",
          "top_keywords"
        ],
        "title": "KeywordResponse"
      },
...
      "MetaInspectRequest": {
        "properties": {
          "html_content": {
            "type": "string",
            "title": "Html Content",
            "example": "<html><head><title>My Great Article Page</title><meta name='description' content='A short description here.'></head><body><h1>Main Title</h1></body></html>"
          }
        },
        "type": "object",
        "required": [
          "html_content"
        ],
        "title": "MetaInspectRequest"
      },
...
      "ReadabilityRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "example": "FastAPI is an asynchronous web framework built on top of Starlette and Pydantic."
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ReadabilityRequest"
      },
      "ReadabilityResponse": {
        "properties": {
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "character_count": {
            "type": "integer",
            "title": "Character Count"
          },
          "estimated_reading_time_minutes": {
            "type": "number",
            "title": "Estimated Reading Time Minutes"
          },
          "reading_difficulty": {
            "type": "string",
            "title": "Reading Difficulty"
          }
        },
        "type": "object",
        "required": [
          "word_count",
          "character_count",
          "estimated_reading_time_minutes",
          "reading_difficulty"
        ],
        "title": "ReadabilityResponse"
      },
...
      "SlugifyRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "example": "10 Best AI & Python Tools for Translators in 2026!"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "default": "en",
            "example": "en"
          },
          "remove_stopwords": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remove Stopwords",
            "default": true,
            "example": true
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SlugifyRequest"
      },
      "SlugifyResponse": {
        "properties": {
          "original_text": {
            "type": "string",
            "title": "Original Text"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          }
        },
        "type": "object",
        "required": [
          "original_text",
          "slug",
          "word_count"
        ],
        "title": "SlugifyResponse"
      },
...
    }
  }
}